+++ sat/cli/hwinv/summary.py
@@ -22,13 +22,17 @@ ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
 OTHER DEALINGS IN THE SOFTWARE.
 """
 from collections import defaultdict
+import logging
 
 from sat.util import format_as_dense_list, get_rst_header, get_pretty_printed_list
+
+LOGGER = logging.getLogger(__name__)
 
 
 class ComponentSummary:
 
-    def __init__(self, comp_type, fields, components, include_xnames):
+    def __init__(self, comp_type, fields, components, include_xnames,
+                 filter_fn=None, display_fields=None, reverse=False):
         """Creates a new ComponentSummary object.
 
         This represents a summary of one component type by multiple fields.
@@ -42,19 +46,38 @@ class ComponentSummary:
                 summarize.
             include_xnames (bool or NoneType): Whether to include xnames in
                 summaries or just counts.
+            filter_fn (dict -> bool): a FilterFunction which is called on
+                each component in the summary to check if it should be
+                included in the summary.
+            display_fields ([str]): the fields which should be summarized
+                in the output.
+            reverse (bool): If True, then the individual summaries will be
+                printed in reverse order (descending.)
         """
         self.comp_type = comp_type
-        self.fields = fields
-        self.components = components
+        self.fields = display_fields or fields
         self.include_xnames = (include_xnames if include_xnames is not None
                                else comp_type.default_show_xnames)
+
+        if filter_fn:
+            try:
+                self.components = [component for component in components
+                                   if filter_fn(component.get_dict(fields, 'canonical_name'))]
+
+            except KeyError as err:
+                LOGGER.warning("Filter key %s does not exist in %s summary. All components will be summarized.",
+                               err, self.comp_type.pretty_name)
+                self.components = components
+        else:
+            self.components = components
 
         self.field_summaries = []
         self.summary_dict = {}
 
         for field in self.fields:
             field_summary = FieldSummary(self.comp_type, field,
-                                         self.components, self.include_xnames)
+                                         self.components, self.include_xnames,
+                                         reverse=reverse)
             self.field_summaries.append(field_summary)
 
     def as_dict(self):
@@ -124,7 +147,7 @@ class ComponentSummary:
 
 class FieldSummary:
 
-    def __init__(self, comp_type, field, components, include_xnames):
+    def __init__(self, comp_type, field, components, include_xnames, reverse=False):
         """Creates a new FieldSummary object.
 
         This represents a summary of one component type by one field.
@@ -138,6 +161,9 @@ class FieldSummary:
                 summarize.
             include_xnames (bool): Whether to include xnames in summaries or
                 just counts.
+            reverse (bool): If True, then the rows will be printed in descending order
+                of the first column when pretty-printed. If False, then they will be
+                printed in ascending order.
         """
         self.comp_type = comp_type
         self.field = field
@@ -162,6 +188,7 @@ class FieldSummary:
 
         # Turn it into a normal dict for ease of dumping with YAML
         self.summary_dict = dict(summary)
+        self.reverse = reverse
 
     def as_dict(self):
         """Gets a dict representation of this field summary.
@@ -175,6 +202,22 @@ class FieldSummary:
         """
         return self.summary_dict
 
+    def prepare_summary(self):
+        """Sorts and possibly reverses the summary contents for printing.
+
+        This is a simple helper function for get_counts_string and
+        get_listings_string.
+
+        Returns:
+            A sequence containing the sorted summary, and if this object's
+            `reverse` attribute is True, then the sorted summary is reversed
+            as well.
+        """
+        summary_items = sorted(self.summary_dict.items())
+        if self.reverse:
+            return reversed(summary_items)
+        return summary_items
+
     def get_counts_string(self):
         """Gets a string representation of the counts in this summary.
 
@@ -184,7 +227,7 @@ class FieldSummary:
 
         table_heading = [self.field.pretty_name, 'Count']
         table_rows = [[attr_value, val_members['count']]
-                      for attr_value, val_members in sorted(self.summary_dict.items())]
+                      for attr_value, val_members in self.prepare_summary()]
         return get_pretty_printed_list(table_rows, table_heading) + '\n'
 
     def get_listings_string(self):
@@ -194,7 +237,7 @@ class FieldSummary:
             A string representation of the listings in this summary.
         """
         result = ''
-        for field_value, val_members in sorted(self.summary_dict.items()):
+        for field_value, val_members in self.prepare_summary():
             result += get_rst_header(
                 '{} {} with {}: {}'.format(val_members['count'],
                                            self.comp_type.plural_pretty_name(),